requirement.txt
ctransformers==0.2.27 
flask==3.0.3 
huggingface_hub==0.26.2
langchain==0.3.7
langchain_community==0.3.7
langchain-core==0.3.19
langchain-huggingface==0.1.2 
langchain-pinecone==0.2.0 
langchain-text-splitters==0.3.2
pinecone-client==5.0.1 
pinecone-notebooks==0.1.1 
pydantic==2.9.2
pypdf==5.1.0 
python-dotenv==1.0.1
requests==2.32.3 
sentence-transformers==3.2.1 

-e .